api_locations/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-firestore==2.14.0
python-dotenv==1.0.1
pyjwt==2.8.0
pydantic>=2.0
pydantic-settings>=2.0

